List the runbooks in the pages site
This PR adds an index page to the pages site. It adds a list of runbooks to the docs/runbooks/README.md file and makes it auto-generated,
so it will add more files to the list if they added to the docs/runbooks directory. The docs/runbooks/README.md file is generated by running
```bash
python automation/generate_pages_index.py
```

This PR also adds a github action to re-generate the docs/runbooks/README.md if the relevant files were changed, and fails the PR if
the file was changed (to force re-generating it by the contributer).

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>