Fix golang module
The monitiringlint is used as a go module, but the repo is not a valid
go module, as the go.mod file is not in the repo's root directory.

This PR moves the go.mod to the root dir, to allow importing the linter
as a module and/or go install the linter cmd.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>